backend/requirements.txt
replicate
pydub
langchain
langchain-google-genai
python-dotenv
requests
Flask
Flask-Cors
protobuf==3.20.3
